+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Use later version of sphinx to generate docs (supports Python 3.10)
+- Update Speechmatics logo
+
 ## [1.0.0] - 2022-03-23
 
 - Publish to pypi.org not test.pypi.org.